# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TestShop API",
        "description": "",
        "contact": {
            "email": "ilyabobkovprog@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Админ"
                ],
                "summary": "Авторизация",
                "operationId": "f3dc2e45e5fc2ee472d07f01f529af03",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "почта",
                                        "type": "string",
                                        "default": "admin@testshop.com"
                                    },
                                    "password": {
                                        "description": "пароль",
                                        "type": "string",
                                        "default": "22vWr!9grE_"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "Админ"
                ],
                "summary": "Выход",
                "operationId": "e72b2aee029d0963731693494aba1b7e",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/category": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Перечень категорий",
                "operationId": "327644f4d836240fa297b2822ab6ca92",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Категории"
                ],
                "summary": "Создать категорию",
                "operationId": "de40ad04180a48262a05cf249f785ab3",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "id категории",
                                        "type": "number",
                                        "default": 1
                                    },
                                    "name": {
                                        "description": "название продукта",
                                        "type": "string",
                                        "default": "Аксессуары"
                                    },
                                    "description": {
                                        "description": "описание",
                                        "type": "string",
                                        "default": "Описание продукта"
                                    },
                                    "image": {
                                        "description": "изображение",
                                        "type": "file"
                                    },
                                    "price": {
                                        "description": "цена",
                                        "type": "number",
                                        "default": 100000
                                    },
                                    "available": {
                                        "description": "в наличии",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ресурс создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Категории"
                ],
                "summary": "Удалить категорию",
                "operationId": "8b7c079d22b38bf71f3189b3b3213731",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id упражнения",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс удален",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Получить категорию",
                "operationId": "25690688339275caf7ded3e8702e7670",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id упражнения",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс получен"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Категории"
                ],
                "summary": "Обновить категорию",
                "operationId": "9be67b4880d28356ade2ebc0b425415e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id упражнения",
                                        "type": "number",
                                        "default": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ресурс обновлен"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Перечень продуктов",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Создать продукт",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "id категории",
                                        "type": "number",
                                        "default": 1
                                    },
                                    "name": {
                                        "description": "название продукта",
                                        "type": "string",
                                        "default": "Аксессуары"
                                    },
                                    "description": {
                                        "description": "описание",
                                        "type": "string",
                                        "default": "Описание продукта"
                                    },
                                    "image": {
                                        "description": "изображение",
                                        "type": "file"
                                    },
                                    "price": {
                                        "description": "цена",
                                        "type": "number",
                                        "default": 100000
                                    },
                                    "available": {
                                        "description": "в наличии",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ресурс создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Получить продукт",
                "operationId": "2768d88590e39a7ded1d0032e64dded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id продукта"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс удалён"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Удалить продукт",
                "operationId": "c830155516bad1f72b0081160edeb2ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id продукта"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс удалён"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Обновить продукт",
                "operationId": "506c53398f4bb15875c04e363a1e9d95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id продукта"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс обновлен"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/shoppingCart": {
            "get": {
                "tags": [
                    "Корзина"
                ],
                "summary": "Перечень продуктов в корзине",
                "operationId": "5cd95b4d85d089b94f26359f2432495c",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Корзина"
                ],
                "summary": "Добавить продукт в корзину",
                "operationId": "05e900d0ca9588733e8b6341157d04ba",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "id продукта",
                                        "type": "number",
                                        "default": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ресурс создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/shoppingCart/{id}": {
            "delete": {
                "tags": [
                    "Корзина"
                ],
                "summary": "Удалить продукт из корзины",
                "operationId": "3a390f4970398b8dfa01e04f4f090477",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id продукта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс удален",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}